requirements.txt
requests
ecdsa
websocket_client
